Permit an "expand" option when accessing the Checks API
When using the Checks API (<https://onfido.com/documentation#checks>), an `expand` option with the value `"reports"` may be supplied in order to return the full reports attached to checks, rather than just their IDs. This is a general feature of the API (see <https://onfido.com/documentation#expanding>) but at the moment it is only implemented for the GET APIs for checks.

This adds support for this expand option.