Drop support for creating Documents and Live Photos from a URL
The gem is currently designed to allow providing a file to upload
when creating a Document or Live Photo in the form of a
`File`-like object or a URL to a file accessible over HTTP. If
the latter is provided, it is fetched over HTTP and put into the
payload for the `POST` request.

However, this functionality as it is currently implemented
represents a security risk, as using `Kernel#open` (monkey-patched
by `open-uri`) can potentially allow executing arbitrary commands
on the server.

This takes the route of deprecating support for passing in URLs.
We will raise an `ArgumentError` with an informative message if
a non-`File` like object is provided (it should respond to `#read`
and `#path`).

We could of course continue to support the existing behaviour,
but:

* It seems perfectly reasonable to expect the caller to safely
fetch the image themselves, and that way, they can do things like
validate its content type, check the URL actually resolves, etc.
* It is actually quite difficult! I spent a fair bit of time
investigating and we would have to re-invent the wheel quite
a lot (e.g. `RestClient` has functionality which allows you to
get a `Tempfile` for a response, but it has the wrong
`#content_type` set ([1][1]) which ends up making the whole thing
a big faff).

[1]: https://github.com/rest-client/rest-client/blob/c2e11b943a8514a94a2b7e68cb0f213cef06a6cc/lib/restclient/request.rb#L783