Remove `Include` setting from Rubocop, which effectively excludes all files
We're currently using `Include` and `Exclude` in our Rubocop
config to turn Rubocop on and off for certain files.

Unfortunately, `Include` overwrites the default list of files,
which is basically *everything*, and means that we're not
actually running Rubocop against most of the library - in fact,
we only run it against the `Rakefile` and `lib/tasks/*.rake`.

This fixes our configuration, removing the `Include`, since those
files will be included by default.